+++ application/view/vectorMesh.h
@@ -9,10 +9,10 @@ public:
 	size_t vertexCount;
 	size_t capacity;
 
-	VectorMesh(const double* vertices, const size_t size) : VectorMesh(vertices, size, size) {};
-	VectorMesh(const double* vertices, const size_t size, size_t capacity);
+	VectorMesh(const float* vertices, const size_t size) : VectorMesh(vertices, size, size) {};
+	VectorMesh(const float* vertices, const size_t size, size_t capacity);
 
-	void update(const double* vertices, const size_t vertexCount);
+	void update(const float* vertices, const size_t vertexCount);
 
 	void render() override;
 	void close() override;
